% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_wrapper.R
\name{fit_fun}
\alias{fit_fun}
\title{A function to fit models with a chosen fitting algorithm. To be used in seq_wrapper.}
\usage{
fit_fun(FUN, ARG)
}
\arguments{
\item{FUN}{Name of a fitting function, like glmmTMB::glmmTMB}

\item{ARG}{A lsit of arguments that can be evaluated by the fitting function}
}
\description{
A function to fit models with a chosen fitting algorithm. To be used in seq_wrapper.
}
