% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_seq_metadata}
\alias{rna_seq_metadata}
\title{Metadata to match the small RNA-seq data set used for illustrations in the vignette and testing.}
\format{
A data frame with 268 rows and 10 variables:
\describe{
  \item{participant}{participant id}
  \item{sex}{mala/female}
  \item{include}{incl if included in analysis}
  \item{rnaseq_include}{incl if included in RNA-seq data set}
  \item{condition}{Multiple, 3 sets per exercise; Single, 1 set per exercise}
  \item{leg}{Right, R; Left, L}
  \item{time}{Study time-point: w0, w2pre, w2post, w12}
  \item{sample.weight}{Sample wet weight (mg)}
  \item{total.rna}{Total amount of RNA (ng)}
  \item{rqi}{RNA quality indicator}
  \item{seq_sample_id}{Sample ID in datasets downloaded with the function download_ome()}}
}
\usage{
rna_seq_metadata
}
\description{
Metadata to match the small RNA-seq data set used for illustrations in the vignette and testing.
}
\keyword{datasets}
