% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_seq_sample}
\alias{rna_seq_sample}
\title{A small (sample) RNA-seq data set used for illustrations in the vignette and testing.}
\format{
## `rna_seq_sample`
A data frame with 1000 rows and 197 columns
\describe{
 \item{transcript_id}{Ensembl transcript id}
 \item{seq_sample_id}{Each remaining column represents sample id}
}
}
\usage{
rna_seq_sample
}
\description{
A small (sample) RNA-seq data set used for illustrations in the vignette and testing.
}
\keyword{datasets}
