% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_wrapper.R
\name{seq_wrapper}
\alias{seq_wrapper}
\title{A flexible upper-level wrapper for iterative modelling using any available fitting algorithm.}
\usage{
seq_wrapper(
  fitting_fun = glmmTMB::glmmTMB,
  arguments,
  data,
  metadata,
  samplename = "seq_sample_id",
  additional_vars = NULL,
  summary_fun = NULL,
  eval_fun = NULL,
  exported = list(),
  subset = NULL,
  cores
)
}
\arguments{
\item{arguments}{A list of arguments to be passed to the fitting function, this should not contain data. Note that the formula must have y as the dependent variable.}

\item{data}{A data frame with targets (i.e. genes, transcripts) as rows and sample names as colums.
The first column is assumed to contain target names/identification}

\item{metadata}{A data frame with sample names (corresponding to column names in the target matrix)
and design variables.}

\item{samplename}{A character value indicating the variable by which metadata can merge with the target data.
This defaults to "seq_sample_id" as this is used in the trainomeMetaData package.}

\item{additional_vars}{A vector of additional variables that is contained in the metadata data set that is needed to fit the model. By default the metadata is reduced to variables contained in the slots formula/model/fixed/random in additional arguments. More variables may be needed for offsets, weights etc.}

\item{summary_fun}{A custom (user-created) function for evaluating/summarizing models. If NULL, a list of fitted models are returned}

\item{eval_fun}{A custom (user-created) function for model diagnostics/evaluation. If NULL, no evaluation/diagnostics of models are returned}

\item{exported}{A list of functions, values etc. to be passed to summary_fun and eval_fun. This list must contain any functions that should be used in model summarise or evaluations.}

\item{subset}{A sequence, random samples or integers to indicate which rows to keep in data. This is useful if you want to test the model in a subset of targets. If keft to the default (NULL), all rows will be used.}

\item{cores}{An integer indicating the number of cores to be used in parallel computations. If NULL, a sequential for loop is used. If "max", all available cores are used.}

\item{fiting_fun}{A model fitting function like stats::lm, glmmTMB::glmmTMB, lme4::lmer}
}
\value{
A nested list with three upper levels slots: models, a list of fitted objects; summaries, a list of summaries created from the summary_fun function; evaluations, a list of diagnostics created from eval_fun.
}
\description{
A flexible upper-level wrapper for iterative modelling using any available fitting algorithm.
}
\details{
This function provides a flexible wrapper to fit, summarize and evaluate statistical models fitted to high dimensional omics-type data.
}
