% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_wrapper.R
\name{transform_merge_fit}
\alias{transform_merge_fit}
\title{Transform, merge and fit models. The function is used inside seq_wrapper to combine
metadata with target-level data and perform the model fitting. The function is used in a
call to pbapply::pblapply.}
\usage{
transform_merge_fit(
  x,
  samp_name = samplename,
  metdat = metadata,
  arg_list = arguments,
  add_vars = additional_vars,
  ffun = fitting_fun
)
}
\arguments{
\item{x}{A list of target-specific counts}

\item{samp_name}{Sample names from the upper level function}

\item{arg_list}{Arguments from the upper level function}

\item{ffun}{the fitting function from the upper level function}
}
\description{
Transform, merge and fit models. The function is used inside seq_wrapper to combine
metadata with target-level data and perform the model fitting. The function is used in a
call to pbapply::pblapply.
}
